feat: refactor admin layout to use server components and add client-side authentication guards
